// engine/generated/materials.ts
export const SCHEMA_VERSION = 3;
// ...
export const MAT = freezeDefinition({
  "AIR": 0,
  "STONE": 1,
  "WATER": 2,
  "SAND": 3,
  "FIRE": 4,
  "METAL": 5,
  "OIL": 6,
  "WOOD": 7,
  "ICE": 8,
  "STEAM": 9,
  "LAVA": 10,
  "GLASS": 11,
  "PLAYER": 12,
  "ASH": 13,
  "SMOKE": 14,
  "OIL_VAPOR": 15,
  "MOLTEN_METAL": 16,
  "MOLTEN_GLASS": 17
} as const);